refactor: renamed file names to snakecase
