+++ convert_gvf_to_vcf/assisting_converter.py
