+++ convert_gvf_to_vcf/etc/dgva_schema.json
