+++ convert_gvf_to_vcf/vcf_line.py
@@ -2,7 +2,7 @@
 The purpose of this file is to populate for each field of a VCF line (and perform any modifications/calculations to achieve this)
 """
 from ebi_eva_common_pyutils.logger import logging_config as log_cfg
-from convert_gvf_to_vcf.assistingconverter import convert_gvf_attributes_to_vcf_values
+from convert_gvf_to_vcf.assisting_converter import convert_gvf_attributes_to_vcf_values
 from dataclasses import dataclass
 from typing import Optional,Union
