+++ tests/test_conversion_statistics.py
@@ -1,9 +1,9 @@
 import unittest
 import os
-from convert_gvf_to_vcf.conversionstatistics import FileStatistics
-from convert_gvf_to_vcf.convertGVFtoVCF import convert_gvf_pragmas_for_vcf_header, generate_vcf_header_line
+from convert_gvf_to_vcf.conversion_statistics import FileStatistics
+from convert_gvf_to_vcf.convert_gvf_to_vcf import convert_gvf_pragmas_for_vcf_header, generate_vcf_header_line
 from convert_gvf_to_vcf.lookup import Lookup
-from convert_gvf_to_vcf.projectpaths import ProjectPaths
+from convert_gvf_to_vcf.project_paths import ProjectPaths
 
 
 class TestConversionStatistics(unittest.TestCase):
