+++ tests/test_convert_gvf_to_vcf.py
@@ -3,12 +3,12 @@ import unittest
 
 
 from convert_gvf_to_vcf.lookup import Lookup
-from convert_gvf_to_vcf.convertGVFtoVCF import generate_vcf_header_unstructured_line, \
+from convert_gvf_to_vcf.convert_gvf_to_vcf import generate_vcf_header_unstructured_line, \
     convert_gvf_pragmas_for_vcf_header, generate_vcf_header_line, parse_pragma, get_pragma_name_and_value, \
     get_pragma_tokens, \
     get_sample_name_from_pragma, get_unique_sample_names, convert_gvf_pragmas_to_vcf_header, \
     convert_gvf_pragma_comment_to_vcf_header, generate_vcf_header_structured_lines, convert
-from convert_gvf_to_vcf.projectpaths import ProjectPaths
+from convert_gvf_to_vcf.project_paths import ProjectPaths
 
 
 class TestConvertGVFtoVCF(unittest.TestCase):
