+++ tests/test_dgva_metadata_retriever.py
@@ -3,9 +3,9 @@ import os
 from unittest import TestCase
 from unittest.mock import patch, MagicMock, Mock, mock_open
 
-from convert_gvf_to_vcf.metadata_retrievers.dgvametadata import DGVAMetadataRetriever
-from convert_gvf_to_vcf.metadata_retrievers.evametadata import EVAMetadataRetriever
-from convert_gvf_to_vcf.projectpaths import ProjectPaths
+from convert_gvf_to_vcf.metadata_retrievers.dgva_metadata import DGVAMetadataRetriever
+from convert_gvf_to_vcf.metadata_retrievers.eva_metadata import EVAMetadataRetriever
+from convert_gvf_to_vcf.project_paths import ProjectPaths
 
 class TestDGVAMetadataRetriever(TestCase):
     def setUp(self):
