+++ tests/test_vcf_line.py
@@ -1,10 +1,10 @@
 import os
 import unittest
 
-from convert_gvf_to_vcf.convertGVFtoVCF import generate_vcf_header_structured_lines
-from convert_gvf_to_vcf.gvffeature import GvfFeatureline
-from convert_gvf_to_vcf.projectpaths import ProjectPaths
-from convert_gvf_to_vcf.vcfline import VcfLine, VcfLineBuilder, VariantRange, extract_reference_allele
+from convert_gvf_to_vcf.convert_gvf_to_vcf import generate_vcf_header_structured_lines
+from convert_gvf_to_vcf.gvf_feature import GvfFeatureline
+from convert_gvf_to_vcf.project_paths import ProjectPaths
+from convert_gvf_to_vcf.vcf_line import VcfLine, VcfLineBuilder, VariantRange, extract_reference_allele
 from convert_gvf_to_vcf.lookup import Lookup
 
 class TestVcfLineBuilder(unittest.TestCase):
